malloc: Take unsigned long as mask to align
Needed to support 64-bit runtime.

+++ src/parisc/malloc.c
@@ -23,7 +23,7 @@ static unsigned long stackptr;
 
 // Allocate physical memory from the given zone and track it as a PMM allocation
 unsigned long
-malloc_palloc(struct zone_s *zone, u32 size, u32 align)
+malloc_palloc(struct zone_s *zone, u32 size, unsigned long align)
 {
     unsigned long data;
 
@@ -42,7 +42,7 @@ malloc_palloc(struct zone_s *zone, u32 size, u32 align)
 
 // Allocate virtual memory from the given zone
 void * __malloc
-parisc_malloc(u32 size, u32 align)
+parisc_malloc(u32 size, unsigned long align)
 {
     return (void*) malloc_palloc(NULL, size, align);
 }
